+++ crates/arf-harp/src/lib.rs
@@ -15,6 +15,6 @@ pub use help::*;
 pub use object::*;
 pub use protect::*;
 pub use startup::{
-    should_ignore_site_r_profile, should_ignore_user_r_profile, source_site_r_profile,
-    source_user_r_profile,
+    call_dot_first, call_dot_first_sys, should_ignore_site_r_profile, should_ignore_user_r_profile,
+    source_site_r_profile, source_user_r_profile,
 };